Transition to Dagger2 successfully!
I have the old functionality back now, working on top of Dagger2.
Definitely more boilerplate, but it will be interesting to see if this
does indeed lead to more comprehensible code. It is nice to not have to
worry about object creation in Activities, that is certainly a plus for
now.