Large refactor. Tabs work
Tabs are set up and working. This is a large refactor to pull all the
logic out of the Activity and push it down into a View that is reusable.
This has some nice factors, I think, although currently it isn't as
slick as it could be with data being destroyed at times when you're
moving between views.

All the tests are broken here as well.